requirements.txt
discord.py==2.6.3
Flask==2.3.3
